Remove IMX related code from OpenMax renderer. Rename to RendererOpenMax
