+++ xbmc/cores/VideoPlayer/VideoRenderers/HwDecRender/Makefile.in
@@ -13,7 +13,7 @@ endif
 ifeq (@USE_OPENGLES@,1)
 SRCS += RendererVTB.cpp
 SRCS += RendererIMX.cpp
-SRCS += RendererOMX.cpp
+SRCS += RendererOpenMax.cpp
 SRCS += RendererMediaCodec.cpp
 SRCS += RendererMediaCodecSurface.cpp
 endif
