+++ xbmc/cores/VideoPlayer/VideoRenderers/HwDecRender/RendererOpenMax.h
